+++ examples/ip_failover/main.tf
@@ -42,7 +42,7 @@ resource "ionoscloud_server" "example_A" {
   cores             = 1
   ram               = 1024
   availability_zone = "ZONE_1"
-  cpu_family        = "AMD_OPTERON"
+  cpu_family        = "INTEL_XEON"
   image_name        = data.ionoscloud_image.example.id
   image_password    = random_password.server_A_image_password.result
   volume {
@@ -78,7 +78,7 @@ resource "ionoscloud_server" "example_B" {
   cores             = 1
   ram               = 1024
   availability_zone = "ZONE_1"
-  cpu_family        = "AMD_OPTERON"
+  cpu_family        = "INTEL_XEON"
   image_name        = data.ionoscloud_image.example.id
   image_password    = random_password.server_B_image_password.result
   volume {