+++ services/clients.go
@@ -12,6 +12,7 @@ import (
 	"github.com/ionos-cloud/terraform-provider-ionoscloud/v6/services/kafka"
 	"github.com/ionos-cloud/terraform-provider-ionoscloud/v6/services/logging"
 	"github.com/ionos-cloud/terraform-provider-ionoscloud/v6/services/nfs"
+	"github.com/ionos-cloud/terraform-provider-ionoscloud/v6/services/vpn"
 )
 
 type SdkBundle struct {
@@ -27,4 +28,5 @@ type SdkBundle struct {
 	LoggingClient      *logging.Client
 	AutoscalingClient  *autoscaling.Client
 	KafkaClient        *kafka.Client
+	VPNClient          *vpn.Client
 }